Catch up the Legate core changes
