+++ cmake/versions.json
@@ -5,7 +5,7 @@
       "git_url" : "https://github.com/magnatelee/legate.core.git",
       "git_shallow": false,
       "always_download": false,
-      "git_tag" : "83eb13edebb0274f0460bd8896b1e84e24fc6de8"
+      "git_tag" : "0a3105b4df9827f5afd2b453c744246635bd7d76"
     }
   }
 }
