+++ src/cunumeric/arg.h
@@ -17,7 +17,6 @@
 #pragma once
 
 #include "legate.h"
-#include "cunumeric/cunumeric_c.h"
 
 namespace cunumeric {
 
@@ -63,8 +62,6 @@ class ArgmaxReduction {
   using RHS = Argval<T>;
 
   static const Argval<T> identity;
-  static const int32_t REDOP_ID =
-    CUNUMERIC_ARGMAX_REDOP * MAX_TYPE_NUMBER + static_cast<int32_t>(legate::legate_type_code_of<T>);
 
   template <bool EXCLUSIVE>
   __CUDA_HD__ inline static void apply(LHS& lhs, RHS rhs)
@@ -85,8 +82,6 @@ class ArgminReduction {
   using RHS = Argval<T>;
 
   static const Argval<T> identity;
-  static const int32_t REDOP_ID =
-    CUNUMERIC_ARGMIN_REDOP * MAX_TYPE_NUMBER + static_cast<int32_t>(legate::legate_type_code_of<T>);
 
   template <bool EXCLUSIVE>
   __CUDA_HD__ inline static void apply(LHS& lhs, RHS rhs)
@@ -100,31 +95,6 @@ class ArgminReduction {
   }
 };
 
-struct register_reduction_op_fn {
-  template <legate::Type::Code CODE, std::enable_if_t<!legate::is_complex<CODE>::value>* = nullptr>
-  void operator()(int32_t type_uid)
-  {
-    using VAL = legate::legate_type_of<CODE>;
-
-    auto runtime = legate::Runtime::get_runtime();
-    auto context = runtime->find_library("cunumeric");
-    {
-      auto redop_id = context->register_reduction_operator<ArgmaxReduction<VAL>>();
-      auto op_kind  = static_cast<int32_t>(legate::ReductionOpKind::MAX);
-      runtime->record_reduction_operator(type_uid, op_kind, redop_id);
-    }
-    {
-      auto redop_id = context->register_reduction_operator<ArgminReduction<VAL>>();
-      auto op_kind  = static_cast<int32_t>(legate::ReductionOpKind::MIN);
-      runtime->record_reduction_operator(type_uid, op_kind, redop_id);
-    }
-  }
-
-  template <legate::Type::Code CODE, std::enable_if_t<legate::is_complex<CODE>::value>* = nullptr>
-  void operator()(int32_t type_uid)
-  {
-    LEGATE_ABORT;
-  }
-};
-
 }  // namespace cunumeric
+
+#include "cunumeric/arg.inl"
