+++ src/cunumeric/unary/unary_op_template.inl
@@ -165,8 +165,8 @@ struct UnaryOpDispatch {
   {
     auto dim = std::max(args.in.dim(), 1);
     if ((OP_CODE == UnaryOpCode::COPY) && (args.in.code() == Type::Code::FIXED_ARRAY)) {
-      auto* type = static_cast<const FixedArrayType*>(args.in.type());
-      cunumeric::double_dispatch(dim, type->num_elements(), UnaryCopyImpl<KIND>{}, args);
+      auto& type = static_cast<const FixedArrayType&>(args.in.type());
+      cunumeric::double_dispatch(dim, type.num_elements(), UnaryCopyImpl<KIND>{}, args);
     } else {
       auto code = OP_CODE == UnaryOpCode::GETARG ? args.out.code() : args.in.code();
       legate::double_dispatch(dim, code, UnaryOpImpl<KIND, OP_CODE>{}, args);
